fix: add RabbitCloudConnectorsAutoConfiguration with @Conditional
