+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-connectors/src/main/java/org/conf/activiti/services/connectors/CloudConnectorsAutoConfiguration.java
@@ -26,43 +26,21 @@ import org.activiti.runtime.api.connector.VariablesMatchHelper;
 import org.activiti.services.connectors.behavior.MQServiceTaskBehavior;
 import org.activiti.services.connectors.message.IntegrationContextMessageBuilderFactory;
 import org.conf.activiti.runtime.api.ConnectorsAutoConfiguration;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.cloud.stream.binder.rabbit.properties.RabbitProducerProperties;
-import org.springframework.cloud.stream.binding.BinderAwareChannelResolver.NewDestinationBindingCallback;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
-import org.springframework.expression.Expression;
-import org.springframework.expression.spel.standard.SpelExpressionParser;
 
 @Configuration
 @AutoConfigureBefore(value = ConnectorsAutoConfiguration.class)
-@PropertySource("classpath:config/integration-result-stream.properties")
 @ComponentScan("org.activiti.core.common.spring.connector")
+@PropertySource("classpath:config/integration-result-stream.properties")
 public class CloudConnectorsAutoConfiguration {
 
-    @Value("${activiti.spring.cloud.stream.connector.integrationRequestSender.routing-key-expression}")
-    private String routingKeyExpression;
-
-    /**
-     * Configures routing key expression for dynamic cloud connector destinations if rabbit binder exists
-     */
-    @Bean
-    @ConditionalOnClass(RabbitProducerProperties.class)
-    @ConditionalOnMissingBean
-    public NewDestinationBindingCallback<RabbitProducerProperties> dynamicConnectorDestinationsBindingCallback() {
-        return (channelName, channel, producerProperties, extendedProducerProperties) -> {
-            Expression expression = new SpelExpressionParser().parseExpression(routingKeyExpression);
-            
-            extendedProducerProperties.setRoutingKeyExpression(expression);
-        };
-    }
     
     @Bean
     @ConditionalOnMissingBean
