+++ activiti-cloud-services-runtime-bundle/activiti-cloud-services-connectors/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,3 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-    org.conf.activiti.services.connectors.CloudConnectorsAutoConfiguration
+    org.conf.activiti.services.connectors.CloudConnectorsAutoConfiguration,\
+	org.conf.activiti.services.connectors.RabbitCloudConnectorsAutoConfiguration
